requirements.txt
Pillow==9.3.0
opencv-python==4.9.0.80
pytesseract==0.3.10